Bot_trading_API_RES/requirements.txt
# Bot Trading API REST Requirements
# Created: 2025-05-22 13:50:39 UTC
# Author: Anhbaza

# Core Dependencies
python-binance>=1.0.17
python-telegram-bot>=13.7
aiohttp>=3.8.1
numpy>=1.21.0
pandas>=1.3.0

# Async Support
asyncio>=3.4.3
aiofiles>=0.8.0

# Utilities
python-dotenv>=0.19.0
PyYAML>=6.0
pytz>=2021.3

# Testing
pytest>=6.2.5
pytest-asyncio>=0.16.0
pytest-cov>=2.12.1

# Logging
colorlog>=6.6.0

# Development
black>=21.12b0
flake8>=4.0.1
mypy>=0.910
isort>=5.10.1

# Documentation
sphinx>=4.3.2
sphinx-rtd-theme>=1.0.0
